+++ ptai-data-structures/src/main/java/com/ptsecurity/appsec/ai/ee/scan/reports/Reports.java
@@ -129,6 +129,7 @@ public class Reports {
             SQL,
             RUBY,
             SOLIDITY,
+            SCALA,
             ALL
         }