+++ ptai-data-structures/src/main/java/com/ptsecurity/appsec/ai/ee/scan/result/ScanBrief.java
@@ -26,7 +26,7 @@ public class ScanBrief {
         @Deprecated V36,
         @Deprecated V40,
         @Deprecated V41,
-        V411, V420, V430, V44X, V450, V460, V470, V471, V472, V480, V481, V490, V491, V4100, V4110, V500;
+        V411, V420, V430, V44X, V450, V460, V470, V471, V472, V480, V481, V490, V491, V4100, V4110, V500, V520;
 
         @SneakyThrows
         public static boolean isDeprecated(@NonNull final ApiVersion version) {
@@ -146,7 +146,8 @@ public class ScanBrief {
             SWIFT("Swift"),
             KOTLIN("Kotlin"),
             RUBY("Ruby"),
-            SOLIDITY("Solidity");
+            SOLIDITY("Solidity"),
+            SCALA("Scala"),;
 
             public static Language fromString(@NonNull final String value) {
                 for (Language language : Language.values())