+++ ptai-data-structures/src/main/java/com/ptsecurity/appsec/ai/ee/scan/settings/AiProjV14ScanSettings.java
@@ -338,6 +338,12 @@ public class AiProjV14ScanSettings extends UnifiedAiProjScanSettings {
     }
 
     @Override
+    public @NonNull Boolean isApplyAllPMRules() {
+        log.trace("No ApplyAllPMRules rules support for AIPROJ schema v.1.4");
+        return false;
+    }
+
+    @Override
     public @NonNull Boolean isUseCustomYaraRules() {
         log.trace("No custom SAST rules support for AIPROJ schema v.1.4");
         return false;