+++ ptai-data-structures/src/main/java/com/ptsecurity/appsec/ai/ee/scan/settings/AiProjV15ScanSettings.java
@@ -348,6 +348,12 @@ public class AiProjV15ScanSettings extends UnifiedAiProjScanSettings {
     }
 
     @Override
+    public @NonNull Boolean isApplyAllPMRules() {
+        log.trace("No ApplyAllPMRules rules support for AIPROJ schema v.1.5");
+        return false;
+    }
+
+    @Override
     public @NonNull Boolean isUseCustomYaraRules() {
         log.trace("No custom SAST rules support for AIPROJ schema v.1.5");
         return false;